Move helper functions
This commit contains no logical changes, it just moves some code to
places where it makes more sense.
